refactor: keep activation split within package budget
